convert.py - remove double check for tokenizer
Remove double for-loop since it's mostly confusing

Also another thing that isn't obvious here, now this will let the file
missing error slip through. The reason for that is that the
tokenizer.json isn't always called tokenizer.json, and the previous error
message did not make it clear that it was infact looking for that file.
So now it's easier for the user to actually see whats going on.